Makes the search bar.

+++ backend/src/post_store_db.ts
@@ -15,13 +15,19 @@ export class PostStoreDB {
   }
 
   async loginUser({ userName, password }: Credentials) {
-    const [isExist] = await this.#users.find({ user: userName }).toArray();
+    const [isExist] = await this.#users
+      .find({ user: userName, password: password })
+      .toArray();
 
     if (isExist !== undefined) {
       return { id: isExist._id };
     }
 
-    const result = await this.#users.insertOne({ user: userName, password });
+    const result = await this.#users.insertOne({
+      user: userName,
+      password: password,
+    });
+
     return { id: result.insertedId.toString() };
   }
 
@@ -42,8 +48,8 @@ export class PostStoreDB {
     return { id: result.insertedId.toString(), date, user };
   }
 
-  async deletePost(id: string) {
-    await this.#posts.deleteOne({ _id: new ObjectId(id) });
+  async deletePost(id: string, userId: string) {
+    await this.#posts.deleteOne({ _id: new ObjectId(id), userId });
 
     return { id };
   }
@@ -56,3 +62,18 @@ export class PostStoreDB {
     return await this.#getAllPostsOfUser(userId);
   }
 }
+
+// const pids = (await this.#posts.find().toArray()).map((_id) => _id);
+// const uids = (await this.#users.find().toArray()).map((_id) => _id);
+// console.log({ pids, uids });
+// pids.map(
+//   async ({ _id }) =>
+//     await this.#posts.deleteOne({ _id: new ObjectId(_id) }),
+// );
+// uids.map(
+//   async ({ _id }) =>
+//     await this.#users.deleteOne({ _id: new ObjectId(_id) }),
+// );
+// const p = (await this.#posts.find().toArray()).map((_id) => _id);
+// const u = (await this.#users.find().toArray()).map((_id) => _id);
+// console.log({ p, u });
